+++ Examples/PHP/ZenKeyOIDCService.php
@@ -15,9 +15,14 @@
  * limitations under the License.
  */
 require __DIR__.'/vendor/autoload.php';
+require __DIR__.'/OIDCProvider.php';
 
 use League\OAuth2\Client\OptionProvider\HttpBasicAuthOptionProvider;
-use League\OAuth2\Client\Provider\GenericProvider;
+use Lcobucci\JWT\Parser;
+use Lcobucci\JWT\ValidationData;
+use Lcobucci\JWT\Signer\Rsa\Sha256;
+use CoderCat\JWKToPEM\JWKConverter;
+
 
 /**
  * This class deals with the ZenKey OAuth2/OpenID Connect flow
@@ -89,7 +94,7 @@ class ZenKeyOIDCService
      *
      * @param string $mccmnc
      *
-     * @return GenericProvider
+     * @return OIDCProvider
      */
     public function discoverOIDCProvider($mccmnc)
     {
@@ -98,13 +103,15 @@ class ZenKeyOIDCService
         $configResponse = file_get_contents($configUrl);
         $openIdConfiguration = json_decode($configResponse);
 
-        $oidcProvider = new GenericProvider([
+        $oidcProvider = new OIDCProvider([
             'clientId' => $this->clientId,
             'clientSecret' => $this->clientSecret,
             'redirectUri' => $this->redirectUri,
             'urlAuthorize' => $openIdConfiguration->authorization_endpoint,
             'urlAccessToken' => $openIdConfiguration->token_endpoint,
             'urlResourceOwnerDetails' => $openIdConfiguration->userinfo_endpoint,
+            'urlJWKs' => $openIdConfiguration->jwks_uri,
+            'issuer' => $openIdConfiguration->issuer
         ]);
 
         // configure the OIDC client to use basic auth with an Authorization header
@@ -137,9 +144,11 @@ class ZenKeyOIDCService
             throw new Exception('state mismatch after carrier discovery');
         }
 
+        $newNonce = random();
         $authorizeParams = [
             'scope' => 'openid', // default to just the basic openid scope
             'login_hint_token' => $loginHintToken,
+            'nonce' => $newNonce
         ];
 
         // if extra params like "context" are needed, pass them in
@@ -156,9 +165,10 @@ class ZenKeyOIDCService
         // generate the URL for the ZenKey authorization endpoint to request an authorization code
         $authorizationUrl = $oidcProvider->getAuthorizationUrl($authorizeParams);
 
-        // persist the mccmnc and a state value in the session
+        // persist the mccmnc, new nonce and a state value in the session
         // for the auth redirect
         $this->sessionService->setState($oidcProvider->getState());
+        $this->sessionService->setNonce($newNonce);
         $this->sessionService->setMCCMNC($mccmnc);
         // TODO: can we remove this without breaking Verizon?
         // remove client_id param because it may break things
@@ -169,15 +179,183 @@ class ZenKeyOIDCService
     }
 
     /**
+     * find a matching key for this JWT from a list of JWKs
+     * 
+     * @param array $keys - array of JWKs
+     * @param Token $jwt
+     * @return array - matching JWK associative array
+     */
+    private function getKeyForJWT($keys, $jwt) {
+        $headerKid = $jwt->getHeader('kid');
+        $headerAlg = $jwt->getHeader('alg');
+
+        // find a matching key in the key list
+        foreach ($keys as $key) {
+            // look for RSA keys in the key list
+            if ($key->kty === 'RSA') {
+                // if the JWT doesn't specific a key ID (kid), use the first matching key
+                // otherwise use the key with the matching kid
+                if (!isset($headerKid) || $key->kid === $headerKid) {
+                    return $key;
+                }
+            } else {
+                // look for keys in the key list with the same algorithm as our JWT
+                // and the same key ID (kid) as our JWT
+                if (isset($key->alg) && $key->alg === $headerAlg && $key->kid === $headerKid) {
+                    return $key;
+                }
+            }
+        }
+        if (isset($headerKid)) {
+            throw new Exception('Unable to find a key for (algorithm, kid):' . $headerAlg . ', ' . $headerKid . ')');
+        }
+
+        throw new Exception('Unable to find a key for RSA');
+    }
+
+    /**
+     * make an HTTP request to get the OIDC provider's JWKs from the jwks_uri
+     * 
+     * @param string $jwksUri
+     * @return array - JSON array object containing a list of JWKs
+     */
+    private function getProviderJWKs($jwksUri) {
+        $jwks = json_decode(file_get_contents($jwksUri));
+        if ($jwks === NULL) {
+            throw new Exception('Error decoding JSON from jwks_uri');
+        }
+        return $jwks->keys;
+    }
+
+    /** 
+     * validate the id token per the spec
+     * https://openid.net/specs/openid-connect-core-1_0.html#IDTokenValidation
+     * 
+     * @param OIDCProvider $oidcProvider
+     * @param string $idToken
+     * @param array $verificationValues - list of values to verify in the ID token
+     * @return void
+     * 
+     * possible $verificationValues
+     * 'accessToken' - verify the at_hash value
+     * 'acrValues' - verify the acr value
+     * 'nonce' - verify the nonce value
+     * 'context' - verify the context value
+     * 'sub' - verify the sub value
+     */
+    private function verifyIdTokenClaims($oidcProvider, $idToken, $verificationValues = []) {
+        // parse the token from a string
+        $decodedIdToken = (new Parser())->parse((string) $idToken);
+
+        // find public keys that this provider uses to sign the JWT
+        $providerJWKs = $this->getProviderJWKs($oidcProvider->getJWKSUrl());
+        // find the specific key used to sign this JWT
+        $jwk = $this->getKeyForJWT($providerJWKs, $decodedIdToken);
+
+        // convert the JWK to PEM format
+        $pem = (new JWKConverter())->toPEM((array) $jwk);
+        // verify the JWT signature
+        $hasValidKey = $decodedIdToken->verify((new Sha256()), $pem);
+
+        // verify the time-based token claims (nbf, iat, exp)
+        // and the issuer and the subject
+        $validationData = new ValidationData();
+        $validationData->setIssuer($oidcProvider->getIssuer());
+        $validationData->setAudience($this->clientId);
+        if(isset($verificationValues['sub'])) {
+            $validationData->setSubject($verificationValues['sub']);
+        }
+        $isValid = $decodedIdToken->validate($validationData);
+
+        if(!$hasValidKey || !$isValid) {
+            throw new Exception('Invalid ID token');
+        }
+
+        // validate nonce if required
+        if (isset($verificationValues['nonce'])
+            && $decodedIdToken->getClaim('nonce') !== $verificationValues['nonce']) {
+            throw new Exception('invalid nonce value in ID token');
+        }
+
+        // verify the at_hash if present
+        if ($decodedIdToken->hasClaim('at_hash')
+            && isset($verificationValues['accessToken'])
+            && !$this->verifyIdTokenAtHash($decodedIdToken, $verificationValues['accessToken'])) {
+            throw new Exception('invalid at_hash value in ID token');
+        }
+
+        // validate context if required
+        if (isset($verificationValues['context'])
+            && $decodedIdToken->hasClaim('context')
+            && $decodedIdToken->getClaim('context') !== $verificationValues['context']) {
+            throw new Exception('invalid context value in ID token');
+        }
+
+        // validate acrValues if required
+        if (isset($verificationValues['acrValues'])
+            && $decodedIdToken->hasClaim('acr')
+            && !in_array($decodedIdToken->getClaim('acr'), explode(' ', $verificationValues['acrValues']))) {
+            throw new Exception('invalid acr value in ID token');
+        }
+    }
+
+    /**
+     * verify the at_hash in the token by hashing the access token and comparing the values
+     * 
+     * @param Token $idToken
+     * @param string $accessToken
+     * @return boolean
+     */
+    private function verifyIdTokenAtHash($idToken, $accessToken) {
+        $idTokenAtHash = $idToken->getClaim('at_hash');
+        $alg = $idToken->getHeader('alg');
+        error_log('accessToken: '.$accessToken);
+        error_log('alg: '.$alg);
+
+        if(isset($alg) && $alg !== 'none'){
+            $bit = substr($alg, 2, 3);
+        } else {
+            $bit = '256';
+        }
+        $len = ((int)$bit)/16;
+        $actualAtHash = $this->urlEncode(substr(hash('sha'.$bit, $accessToken, true), 0, $len));
+
+        if ($idTokenAtHash != $actualAtHash) {
+            error_log('at_hash: '.$idTokenAtHash.' :: actualhash: '.$actualAtHash);
+            return false;
+        }
+        return true;
+    }
+
+    /**
+     * Base64 URL encode a value
+     * @param string $str
+     * @return string
+     */
+    protected function urlEncode($str) {
+        $enc = base64_encode($str);
+        $enc = rtrim($enc, '=');
+        $enc = strtr($enc, '+/', '-_');
+        return $enc;
+    }
+
+    /**
      * We have an auth code, we can now exchange it for a token.
      *
      * @param GenericProvider $oidcProvider
      * @param string          $code
      * @param string          $state
+     * @param array           $tokenVerificationValues - list of values to verify in the ID token
+     * 
+     * possible $tokenVerificationValues
+     * 'acrValues' - verify the acr value
+     * 'nonce' - verify the nonce value
+     * 'context' - verify the context value
+     * 'sub' - verify the sub value
      *
      * @return League\OAuth2\Client\Token\AccessTokenInterface
      */
-    public function requestToken($oidcProvider, $code, $state)
+    public function requestToken($oidcProvider, $code, $state, $tokenVerificationValues = [])
     {
         // prevent request forgeries by checking that the incoming state matches
         if ($state !== $this->sessionService->getState()) {
@@ -188,6 +366,15 @@ class ZenKeyOIDCService
         $tokens = $oidcProvider->getAccessToken('authorization_code', [
             'code' => $code,
         ]);
+
+        // make sure to verify the nonce
+        $tokenVerificationValues['nonce'] = $this->sessionService->getNonce();
+
+        // verify the tokens
+        $this->verifyIdTokenClaims(
+            $oidcProvider,
+            $tokens->getValues()['id_token'],
+            array_merge(['accessToken' => $tokens->getToken()], $tokenVerificationValues));
 
         // clear the state
         $this->sessionService->clearState();