+++ Examples/Python/CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 ## Versions
+- [2020-03-27](#2020-03-27)
 - [2020-02-27](#2020-02-27)
 - [2020-01-31](#2020-01-31)
 
@@ -24,7 +25,13 @@ When the Unreleased section becomes a new version, duplicate the Template to cre
 ```
 
 ## Unreleased
-- none
+None
+
+## 2020-03-27
+### Changed
+- Updated support email address in Readme
+- Use pipenv and commit Pipfile and Pipfile.lock
+- Send a nonce in the authorization request and verify it in the ID token. This helps prevent replay attacks
 
 ## 2020-02-27
 ### Added